@@ -1,6 +1,6 @@
 class DeleteDatabase < SireneAsAPIInteractor
   def call
     stdout_info_log('Deleting database...')
-    Etablissement.delete_all
+    EtablissementV2.delete_all
   end
 end
