@@ -45,7 +45,7 @@ class ImportMonthlyStockCsv < SireneAsAPIInteractor
   end
 
   def clean_database?
-    Etablissement.first.nil?
+    EtablissementV2.first.nil?
   end
 
   def create_progressbar(context)
@@ -66,10 +66,7 @@ class ImportMonthlyStockCsv < SireneAsAPIInteractor
     {
       chunk_size: 2_000,
       col_sep: ',',
-      row_sep: "\n",
-      convert_values_to_numeric: false,
-      key_mapping: {},
-      file_encoding: 'UTF-8'
+      row_sep: "\n"
     }
   end
 
@@ -85,9 +82,9 @@ class ImportMonthlyStockCsv < SireneAsAPIInteractor
   end
 
   def stdout_etablissement_count_change
-    etablissement_count_before = Etablissement.count
+    etablissement_count_before = EtablissementV2.count
     yield
-    etablissement_count_after = Etablissement.count
+    etablissement_count_after = EtablissementV2.count
 
     entries_added = etablissement_count_after - etablissement_count_before
 
