@@ -1,5 +1,6 @@
 class UpdateEtablissementRowsJob < EtablissementRowJobs
   attr_accessor :lines
+
   def initialize(lines)
     @lines = lines
   end
@@ -9,8 +10,8 @@ class UpdateEtablissementRowsJob < EtablissementRowJobs
 
     begin
       update_attrs(etablissements)
-    rescue StandardError => error
-      stdout_error_log "Error: Could not update etablissement attributes:  #{error.class}
+    rescue StandardError => e
+      stdout_error_log "Error: Could not update etablissement attributes:  #{e.class}
         Make sure Solr server is launched on the right environment and accessible."
       exit
     end
@@ -18,11 +19,12 @@ class UpdateEtablissementRowsJob < EtablissementRowJobs
 
   def update_attrs(etablissements)
     etablissements.each do |etablissement_attrs|
-      etablissement = Etablissement.find_or_initialize_by(siret: etablissement_attrs[:siret])
+      etablissement = EtablissementV2.find_or_initialize_by(siret: etablissement_attrs[:siret])
 
       nature_mise_a_jour = etablissement_attrs[:nature_mise_a_jour]
       # Il y a une paire I/F, I etant l'etat initial
       next if nature_mise_a_jour == 'I'
+
       if nature_mise_a_jour == 'E'
         # On supprime l'etablissement si il est persiste
         # Si non persiste veut dire qu on rejoue un patch interrompu
