@@ -1,3 +1,4 @@
+# rubocop:disable Style/ClassAndModuleChildren
 module PayloadSerializer
   class SirenPayload
     def initialize(siren, result_siege, results_sirets)
@@ -67,6 +68,7 @@ module PayloadSerializer
 
     def status
       return 200 unless @result_siege.nil?
+
       404
     end
 
@@ -74,7 +76,7 @@ module PayloadSerializer
     def date_sirene_stock
       name_stock = File.read(SaveLastMonthlyStockName.new.full_path)
       name_stock.split('/')[4]
-    rescue
+    rescue StandardError
       nil
     end
 
@@ -85,3 +87,4 @@ module PayloadSerializer
     end
   end
 end
+# rubocop:enable Style/ClassAndModuleChildren
