@@ -5,7 +5,7 @@ require 'json'
 class SolrRequests < SireneAsAPIInteractor
   attr_accessor :keyword
 
-  def initialize *keywords
+  def initialize(*keywords)
     keyword = keywords[0].to_s.gsub(/[+<>'"=&,;\n]/, ' ') # Get first word in params & Prevent Solr injections
     @keyword = URI.encode_www_form_component(keyword)
   end
@@ -19,8 +19,8 @@ class SolrRequests < SireneAsAPIInteractor
     stdout_info_log 'Building suggester dictionary... This might take a while (~30 mins)'
     begin
       request_build_dictionary
-    rescue StandardError => error
-      stdout_error_log "Error while building dictionary : #{error}"
+    rescue StandardError => e
+      stdout_error_log "Error while building dictionary : #{e}"
     else
       stdout_success_log('Dictionary was correctly built !')
     end
@@ -29,19 +29,26 @@ class SolrRequests < SireneAsAPIInteractor
   private
 
   def uri_solr
-    "/solr/#{Rails.env}/suggesthandler?wt=json&suggest.q=#{@keyword}"
+    "#{solr_base_url}/suggesthandler?wt=json&suggest.q=#{@keyword}"
   end
 
   def request_build_dictionary
     http_session = Net::HTTP.new('localhost', solr_port)
-    # http_session.read_timeout = 14_400 # 4 hours max to build dictionary
+    # http_session.read_timeout = 14_400 # 10 hours max to build dictionary
     http_session.read_timeout = 60 * 60 * 10 # 10 hours
-    uri = "/solr/#{Rails.env}/suggesthandler?suggest.build=true"
+    uri = "#{solr_base_url}/suggesthandler?suggest.build=true"
     http_session.get(uri)
   end
 
   def solr_port
-    sunspot_config = YAML.load_file('config/sunspot.yml')
-    sunspot_config[Rails.env]['solr']['port']
+    solr_conf['port']
+  end
+
+  def solr_base_url
+    solr_conf['path']
+  end
+
+  def solr_conf
+    Rails.application.config_for(:sunspot)['solr']
   end
 end
