@@ -1,4 +1,8 @@
+require 'sidekiq/web'
+require 'sidekiq/cron/web'
+
 Rails.application.routes.draw do
+  mount Sidekiq::Web => '/sidekiq'
 
   concern :v1_routes do
     get 'siret/:siret' => '/api/v1/siret#show', :constraints => { :siret => /.*/ }
@@ -20,12 +24,24 @@ Rails.application.routes.draw do
     get 'siren/:siren/etablissements_geojson' => '/api/v2/siren_children_geojson#show'
   end
 
+  concern :v3_routes do
+    get 'unites_legales/' => '/api/v3/unites_legales#index'
+    get 'unites_legales/:siren' => '/api/v3/unites_legales#show'
+
+    get 'etablissements/' => '/api/v3/etablissements#index'
+    get 'etablissements/:siret' => '/api/v3/etablissements#show'
+  end
+
   namespace :v1 do
     concerns :v1_routes
   end
 
   namespace :v2 do
     concerns :v2_routes
+  end
+
+  namespace :v3 do
+    concerns :v3_routes
   end
 
   # DIRTY FIX (nginx configuration and url prefixing do not work)
@@ -37,6 +53,9 @@ Rails.application.routes.draw do
     namespace :v2 do
       concerns :v2_routes
     end
-  end
 
+    namespace :v3 do
+      concerns :v3_routes
+    end
+  end
 end
