@@ -4,7 +4,7 @@ describe API::V1::NearEtablissementController do
   context 'when there are results', type: :request do
     let!(:etablissement_to_search) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 1,
         siret: '123456',
         latitude: '48.000001',
@@ -14,7 +14,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_find) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 2,
         siret: '123457',
         latitude: '48.000002',
@@ -24,7 +24,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_not_find) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 3,
         siret: '123458',
         latitude: '50.000000', # Unnamed road in Kazakhstan
@@ -33,7 +33,7 @@ describe API::V1::NearEtablissementController do
       )
     end
     it 'return results in 200 payload' do
-      Etablissement.reindex
+      EtablissementV2.reindex
 
       get '/v1/near_etablissement/123456'
 
@@ -53,7 +53,7 @@ describe API::V1::NearEtablissementController do
   context 'when no results', type: :request do
     let!(:etablissement_to_search) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 1,
         siret: '123456',
         latitude: '48.000001',
@@ -63,7 +63,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_not_find) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 3,
         siret: '123458',
         latitude: '50.000000',
@@ -72,7 +72,7 @@ describe API::V1::NearEtablissementController do
       )
     end
     it 'returns 404 payload' do
-      Etablissement.reindex
+      EtablissementV2.reindex
 
       get '/v1/near_etablissement/123456'
 
@@ -83,9 +83,9 @@ describe API::V1::NearEtablissementController do
   end
 
   context 'when SIRET isnt found', type: :request do
-    let!(:etablissement_to_not_find) { create(:etablissement, id: 1, siret: '123456') }
+    let!(:etablissement_to_not_find) { create(:etablissement_v2, id: 1, siret: '123456') }
     it 'returns 400 payload' do
-      Etablissement.reindex
+      EtablissementV2.reindex
 
       get '/v1/near_etablissement/999999'
 
@@ -98,7 +98,7 @@ describe API::V1::NearEtablissementController do
   context 'when there are etablissements with other activity', type: :request do
     let!(:etablissement_to_search) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 1,
         siret: '123456',
         latitude: '48.000001',
@@ -108,7 +108,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_find) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 2,
         siret: '123457',
         latitude: '48.000002',
@@ -118,7 +118,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_find2) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 3,
         siret: '123458',
         latitude: '48.000001',
@@ -127,7 +127,7 @@ describe API::V1::NearEtablissementController do
       )
     end
     it 'find them' do
-      Etablissement.reindex
+      EtablissementV2.reindex
 
       get '/v1/near_etablissement/123456'
 
@@ -142,7 +142,7 @@ describe API::V1::NearEtablissementController do
   context 'when there are etablissements with other activity', type: :request do
     let!(:etablissement_to_search) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 1,
         siret: '123456',
         latitude: '48.000001',
@@ -152,7 +152,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_find) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 2,
         siret: '123457',
         latitude: '48.000001',
@@ -162,7 +162,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_not_find) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 3,
         siret: '123458',
         latitude: '48.000001',
@@ -171,7 +171,7 @@ describe API::V1::NearEtablissementController do
       )
     end
     it 'doesnt find them if user doesnt want to' do
-      Etablissement.reindex
+      EtablissementV2.reindex
 
       get '/v1/near_etablissement/123456?only_same_activity=true'
 
@@ -187,7 +187,7 @@ describe API::V1::NearEtablissementController do
   context 'when there are etablissements with other approximate activity', type: :request do
     let!(:etablissement_to_search) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 1,
         siret: '123456',
         latitude: '48.000001',
@@ -197,7 +197,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_find) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 2,
         siret: '123457',
         latitude: '48.000001',
@@ -207,7 +207,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_not_find) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 3,
         siret: '123458',
         latitude: '48.000001',
@@ -216,7 +216,7 @@ describe API::V1::NearEtablissementController do
       )
     end
     it 'find only them if user want to' do
-      Etablissement.reindex
+      EtablissementV2.reindex
 
       get '/v1/near_etablissement/123456?approximate_activity=true'
 
@@ -232,7 +232,7 @@ describe API::V1::NearEtablissementController do
   context 'when specifying the radius', type: :request do
     let!(:etablissement_to_search) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 1,
         siret: '123456',
         latitude: '48.000001',
@@ -242,7 +242,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_find) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 2,
         siret: '123457',
         latitude: '48.000009',
@@ -252,7 +252,7 @@ describe API::V1::NearEtablissementController do
     end
     let!(:etablissement_to_find_after) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 3,
         siret: '123458',
         latitude: '49.00000', # Unnamed road in Kazakhstan
@@ -261,7 +261,7 @@ describe API::V1::NearEtablissementController do
       )
     end
     it 'correctly find the results' do
-      Etablissement.reindex
+      EtablissementV2.reindex
 
       get '/v1/near_etablissement/123456'
 
@@ -277,7 +277,7 @@ describe API::V1::NearEtablissementController do
   context 'when using param per_page', type: :request do
     let!(:etablissement_to_search) do
       create(
-        :etablissement,
+        :etablissement_v2,
         id: 1,
         siret: '123456',
         latitude: '48.000001',
@@ -287,7 +287,7 @@ describe API::V1::NearEtablissementController do
     end
     it 'works' do
       populate_with_11_local_companies('48.000001', '3.000001')
-      Etablissement.reindex
+      EtablissementV2.reindex
 
       get '/v1/near_etablissement/123456'
       result_hash = body_as_json
