@@ -1,14 +1,14 @@
-# TODO: Check validation siren dans github SIAD
-# Rappel : Mise a jour => Clef I/F
-# Creation : C, Entree diffusion : D
-# Suppression : E, Sortie diffusion : O
-
 FactoryBot.define do
-  factory :etablissement do
-    sequence(:nom) { |n| "entreprise#{n}" }
-    sequence(:siren) { |n| "11111#{n}" }
-    sequence(:date_mise_a_jour, ('00'..'59').cycle) { |n| "2017-01-01T10:25:#{n}" }
-    nature_mise_a_jour %w[I F C D].sample
-    code_postal %w[92600 75015 34000 93000 55055 07300].cycle
+  factory :etablissement do |_etablissement|
+    sequence(:enseigne_1) { |n| "enseigne_#{n}" }
+    sequence(:siren) { |n| "00000000#{n}" }
+    sequence(:nic) { |n| "0000#{n}" }
+    siret { siren + nic }
+    statut_diffusion { 'O' }
+    date_dernier_traitement { '2016-11-30T14:31:01' }
+
+    trait :non_diffusable do
+      statut_diffusion { 'N' }
+    end
   end
 end
