@@ -1,7 +1,9 @@
 require 'rails_helper'
 
 describe GetLastMonthlyStockLink do
-  subject(:context) { described_class.call}
+  include_context 'mute interactors'
+
+  subject(:context) { described_class.call }
 
   context 'when file server is down', vcr: { cassette_name: 'geo-sirene_server_KO' } do
     it 'fails' do
@@ -9,19 +11,17 @@ describe GetLastMonthlyStockLink do
     end
   end
 
-  # HOTFIX july 19: API was changed to always return march (03) file. cassette_name changed from geo-sirene_march
-  # to geo-sirene_post_july
-  context 'when in march', vcr: { cassette_name: 'geo-sirene_post_july' } do
+  context 'when in march', vcr: { cassette_name: 'geo-sirene_march' } do
     it 'retrieves february stock link' do
       allow(Time).to receive(:now) { Time.new(2018, 3, 31) }
 
-      # last_montly_link = 'http://data.cquest.org/geo_sirene/2018-02/geo_sirene.csv.gz'
-      last_montly_link = 'http://data.cquest.org/geo_sirene/2019-03/geo_sirene.csv.gz'
+      last_montly_link = 'http://data.cquest.org/geo_sirene/2018-02/etablissements_actifs.csv.gz'
       expect(described_class.call.link).to eq(last_montly_link)
     end
   end
 
-  # context 'when in march but link is not available', vcr: { cassette_name: 'geo-sirene_march_link_not_available' } do
+  # context 'when in march but link is not available',
+  # vcr: { cassette_name: 'geo-sirene_march_link_not_available' } do
   #   it 'fails' do
   #     allow(Time).to receive(:now) { Time.new(2018, 3, 31) }
 
@@ -29,7 +29,8 @@ describe GetLastMonthlyStockLink do
   #   end
   # end
 
-  # context 'When on the 1st of april and no stock link yet', vcr: { cassette_name: 'geo-sirene_1_april_no_stock_yet' } do
+  # context 'When on the 1st of april and no stock link yet',
+  # vcr: { cassette_name: 'geo-sirene_1_april_no_stock_yet' } do
   #   it 'retrieves february stock link' do
   #     allow(Time).to receive(:now) { Time.new(2018, 4, 1) }
 
