@@ -1,10 +1,11 @@
 require 'rails_helper'
 
 describe GetRelevantPatchesLinks do
-  context 'when in march & there are LESS than 5 patches since last monthly update,',
-    vcr: { cassette_name: 'geo-sirene_file_index_4_links_since_LM' } do
+  include_context 'mute interactors'
 
-    let!(:etablissement) { create(:etablissement, date_mise_a_jour: '2017-03-02T10:55:43') }
+  context 'when in march & there are LESS than 5 patches since last monthly update,',
+          vcr: { cassette_name: 'geo-sirene_file_index_4_links_since_LM' } do
+    let!(:etablissement) { create(:etablissement_v2, date_mise_a_jour: '2017-03-02T10:55:43') }
 
     it 'retrieves the right number of patches' do
       allow(File).to receive(:read) { 'http://data.cquest.org/geo_sirene/2017-02/geo-sirene.csv.gz' }
@@ -17,9 +18,8 @@ describe GetRelevantPatchesLinks do
 
   context 'when in march & there are MORE than 5 patches since last monthly update
     & there are LESS than 5 patches since last update,',
-    vcr: { cassette_name: 'geo-sirene_file_index_20170330_22_links_since_LM' } do
-
-    let!(:etablissement) { create(:etablissement, date_mise_a_jour: '2017-03-28T10:55:43') }
+          vcr: { cassette_name: 'geo-sirene_file_index_20170330_22_links_since_LM' } do
+    let!(:etablissement) { create(:etablissement_v2, date_mise_a_jour: '2017-03-28T10:55:43') }
 
     it 'retrieves 5 patches' do
       allow(File).to receive(:read) { 'http://data.cquest.org/geo_sirene/2017-02/geo-sirene.csv.gz' }
@@ -32,9 +32,8 @@ describe GetRelevantPatchesLinks do
 
   context 'when in march & there are MORE than 5 patches since last monthly update
     & MORE than 5 since last daily update,',
-    vcr: { cassette_name: 'geo-sirene_file_index_20170330_22_links_since_LM' } do
-
-    let!(:etablissement) { create(:etablissement, date_mise_a_jour: '2017-03-15T10:55:43') }
+          vcr: { cassette_name: 'geo-sirene_file_index_20170330_22_links_since_LM' } do
+    let!(:etablissement) { create(:etablissement_v2, date_mise_a_jour: '2017-03-15T10:55:43') }
 
     it 'retrieves the right number of patches' do
       allow(File).to receive(:read) { 'http://data.cquest.org/geo_sirene/2017-02/geo-sirene.csv.gz' }
@@ -46,7 +45,7 @@ describe GetRelevantPatchesLinks do
   end
 
   context 'when rebuilding database in the middle of month', vcr: { cassette_name: 'geo-sirene_20181215' } do
-    let!(:etablissement) { create :etablissement, date_mise_a_jour: '2018-11-30T20:02:32' }
+    let!(:etablissement) { create :etablissement_v2, date_mise_a_jour: '2018-11-30T20:02:32' }
 
     it 'retrieves all the patches since the begining of the month' do
       allow(File).to receive(:read) { 'http://data.cquest.org/geo_sirene/2018-11/geo_sirene.csv.gz' }
@@ -58,9 +57,8 @@ describe GetRelevantPatchesLinks do
   end
 
   context 'when in january and it is the first patches of the new year',
-    vcr: { cassette_name: 'geo-sirene_file_index_20190102' } do
-
-    let!(:etablissement) { create :etablissement, date_mise_a_jour: '2018-12-31T20:02:32' }
+          vcr: { cassette_name: 'geo-sirene_file_index_20190102' } do
+    let!(:etablissement) { create :etablissement_v2, date_mise_a_jour: '2018-12-31T20:02:32' }
 
     it 'retrieves only one patch' do
       allow(File).to receive(:read) { 'http://data.cquest.org/geo_sirene/2018-12/geo_sirene.csv.gz' }
