@@ -1,4 +1,4 @@
-# Config for factory_girl to work with RSpec
+# Config for factory_bot to work with RSpec
 RSpec.configure do |config|
   config.include FactoryBot::Syntax::Methods
 end
